+++ back-end/api/app/Models/Appliance.php
@@ -9,8 +9,11 @@ class Appliance extends Model
 {
     use HasFactory;
 
+    protected $guarded = [
+        'id'
+    ];
+
     protected $fillable = [
-        'id',
         'name',
         'description',
         'voltage',