+++ front-end/app/src/main.ts
@@ -1,11 +1,14 @@
 import './assets/main.css'
 
-import { createApp } from 'vue'
+import { createApp, defineAsyncComponent } from 'vue'
 import App from './App.vue'
 import router from './router'
+import store from './store'
+import FormView from "./components/Form/FormView.vue"
 
 const app = createApp(App)
 
 app.use(router)
+app.use(store)
 
 app.mount('#app')