+++ front-end/app/src/utils/Form/interfaces/IForm.ts
@@ -1,0 +1,8 @@
+interface IForm{
+    type : "create" | "update",
+    id : number,
+    name : string,
+    description : string,
+    brand : string,
+    voltage : string
+}